requirements.txt
fastapi
python-jose
uvicorn[standard]